requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
openai>=1.30.0
pymupdf>=1.24.0
pymupdf4llm>=0.0.17
httpx>=0.27.0
pyyaml>=6.0
pydantic>=2.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
upstash-redis>=1.0.0
